[tpcgen-cli]: split CLI integration tests
